pyproject.toml
[project]
name = "probarb"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "aiohttp>=3.13.2",
    "dotenv>=0.9.9",
    "fastapi>=0.121.3",
    "httpx>=0.28.1",
    "matplotlib>=3.10.7",
    "mysql-connector-python>=9.5.0",
    "nest-asyncio>=1.6.0",
    "pandas>=2.3.3",
    "playwright>=1.55.0",
    "py-clob-client>=0.28.0",
    "pyaml>=25.7.0",
    "pytest>=8.4.2",
    "pytest-asyncio>=1.3.0",
    "pyyaml>=6.0.3",
    "requests>=2.32.5",
    "rich>=14.2.0",
    "web3>=7.14.0",
    "websockets>=15.0.1",
]
